Remove spantree and range modules + Cleanup

+++ srcs/editor/binding/delete.c
@@ -6,7 +6,7 @@
 /*   By: juloo <juloo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/12 22:00:44 by juloo             #+#    #+#             */
-/*   Updated: 2016/01/26 14:37:15 by jaguillo         ###   ########.fr       */
+/*   Updated: 2016/01/26 19:45:10 by jaguillo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ bool			editor_bind_delete(t_editor *editor, uint32_t flags)
 	if (editor->sel != 0)
 		move = editor->sel;
 	else if ((move = cursor_move(editor, flags)) == 0
-		|| (editor->cursor + move) < 0
+		|| ((int)editor->cursor + move) < 0
 		|| (editor->cursor + move) > editor->text.length)
 		return (false);
 	if (flags & DELETE_CLIPBOARD)
